Move syslog copy to internal
since we don't want to expose it.
