+++ cmd/internal/syslog/syslog.go
