+++ cmd/internal/syslog/syslog_plan9.go
