+++ cmd/internal/syslog/syslog_test.go
