+++ cmd/internal/syslog/syslog_unix.go
