+++ cmd/internal/syslog/syslog_windows.go
